requirements.txt
dask==2023.4.1
google-cloud-bigquery==3.10.0
google-cloud-storage==2.9.0
pandas==2.0.1
pyarrow==12.0.0
pyspark~=3.4.0
numpy~=1.24.3